refactor: rename settings.py to _settings.py and update imports accordingly
