+++ packages/kiarina-lib-cloudflare-d1/src/kiarina/lib/cloudflare/d1/_core/models/d1_context.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 
 from kiarina.lib.cloudflare.auth import CloudflareAuthSettings
 
-from ...settings import D1Settings
+from ..._settings import D1Settings
 
 
 @dataclass
