+++ packages/kiarina-lib-cloudflare-d1/src/kiarina/lib/cloudflare/d1/_sync/models/d1_client.py
@@ -5,7 +5,7 @@ from kiarina.lib.cloudflare.auth import CloudflareAuthSettings
 from ..._core.models.d1_context import D1Context
 from ..._core.operations.query import query
 from ..._core.views.result import Result
-from ...settings import D1Settings
+from ..._settings import D1Settings
 
 
 class D1Client:
