+++ packages/kiarina-lib-cloudflare-d1/src/kiarina/lib/cloudflare/d1/asyncio.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from ._async.helpers.create_d1_client import create_d1_client
     from ._async.models.d1_client import D1Client
-    from .settings import D1Settings, settings_manager
+    from ._settings import D1Settings, settings_manager
 
 __all__ = [
     # ._async.helpers
